pyproject.toml
[project]
name = "cad-widgets"
version = "0.1.0"
description = "OpenCascade PySide6 3D CAD viewer widgets with modular architecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pySide6>=6.6.0",
    "cadquery-ocp>=7.7.0",
    "numpy>=1.24.0",
]
authors = [
    {name = "CAD Widgets Contributors"}
]
keywords = ["cad", "opencascade", "pyside6", "3d", "viewer", "widget"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/yourusername/cad-widgets"
Repository = "https://github.com/yourusername/cad-widgets"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/cad_widgets"]

[dependency-groups]
dev = [
    "pytest>=8.3.5,<9",
    "pytest-datadir>=1.7.2,<2",
    "invoke>=2.2.0,<3",
    "ruff>=0.15.2",
    "pytest-cov>=7.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
